Changes to home page. Fixes #56
