+++ lib/cubit/index_cubit.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class IndexCubit extends Cubit<int> {
+  IndexCubit() : super(0);
+
+  void changeIndex(int newIndex) {
+    emit(newIndex);
+  }
+}
