+++ lib/pages/explore.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+// Explore Page
+// All the modules will be placed here like alphabets, animals, etc...
+// TODO: Implement the Explore Page
+
+class ExplorePage extends StatelessWidget {
+  const ExplorePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Explore Page"),
+      )
+    );
+  }
+}
