+++ lib/pages/favorite.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class FavoritePage extends StatelessWidget {
+  const FavoritePage({super.key});
+
+  // Favorite Page
+  // All the favorited modules by the user will be placed here
+  // TODO: Implement the Favorite Page
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text("Favorited Items here"),
+    );
+  }
+}
