+++ lib/pages/modules/solar.dart
