+++ pubspec.yaml
@@ -43,6 +43,8 @@ dependencies:
   flutter_card_swiper: ^7.0.0
   adaptive_theme: ^3.6.0
   fluttertoast: ^8.2.5
+  google_nav_bar: ^5.0.6
+  flutter_bloc: ^8.1.5
 
 dev_dependencies:
   flutter_test:
